# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CAFA Program Management Information System (PMIS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get the current user's full profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401":
          description: Not authenticated
    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update the current user's editable profile fields
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProfileInput" }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "400":
          description: Validation error
        "422":
          description: Invalid notification preference settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidationErrorResponse" }
        "401":
          description: Not authenticated
  /profile/change-password:
    post:
      operationId: changePassword
      tags: [profile]
      summary: Change the current user's password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordInput" }
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
        "400":
          description: Weak password or validation error
        "401":
          description: Current password incorrect or not authenticated
        "429":
          description: Too many password-change attempts
  /profile/photo/upload-url:
    post:
      operationId: requestProfilePhotoUploadUrl
      tags: [profile]
      summary: Request a short-lived upload URL for the signed-in user's profile photo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfilePhotoUploadRequest" }
      responses:
        "200":
          description: Short-lived private upload descriptor
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfilePhotoUploadDescriptor" }
        "413":
          description: Photo exceeds the 5 MB limit
        "415":
          description: Unsupported photo type
  /profile/photo:
    get:
      operationId: getProfilePhoto
      tags: [profile]
      summary: Stream the signed-in user's current profile photo through an authenticated proxy
      responses:
        "200":
          description: Profile image
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        "404":
          description: No profile photo
    post:
      operationId: completeProfilePhotoUpload
      tags: [profile]
      summary: Verify and attach an uploaded photo to the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfilePhotoCompleteInput" }
      responses:
        "200":
          description: Photo attached
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfilePhotoResponse" }
        "400":
          description: Invalid or expired upload descriptor
        "413":
          description: Actual uploaded size exceeds the limit
        "415":
          description: Uploaded content is not a permitted image
    delete:
      operationId: removeProfilePhoto
      tags: [profile]
      summary: Remove the signed-in user's profile photo
      responses:
        "200":
          description: Photo removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfilePhotoResponse" }
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, role, scope]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, description: "super_admin | executive_director | program_manager | senior_program_coordinator | technical_coordinator | state_office_manager | state_program_officer | viewer" }
        roleLabel: { type: string }
        scope: { type: string, description: "hq | state" }
        stateId: { type: ["integer", "null"] }
        stateName: { type: ["string", "null"] }
        stateNameAr: { type: ["string", "null"], description: "Canonical Arabic State label for locale-aware display." }
        sector: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
...
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            username: { type: ["string", "null"] }
            phone: { type: ["string", "null"] }
            status: { type: string, description: "active | invited | suspended | inactive | deactivated" }
            languagePreference: { type: string, description: "en | ar" }
            lastLoginAt: { type: ["string", "null"], format: date-time }
            isOnline: { type: boolean, description: "True only while this API process has an authenticated Socket.IO connection for the user." }
            lastSeenAt: { type: ["string", "null"], format: date-time, description: "Persisted timestamp of the last final realtime-presence transition to offline." }
            createdAt: { type: ["string", "null"], format: date-time }
            updatedAt: { type: ["string", "null"], format: date-time }
            hasInvite: { type: boolean, description: "True when an unaccepted invite token exists. The token value itself is only returned once at issuance." }
            emailVerified: { type: boolean, description: "True if the user's email address has been verified." }
            emailVerifiedAt: { type: ["string", "null"], format: date-time, description: "When the email was verified." }
...
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/UserDetail"
        - type: object
          properties:
            jobTitle: { type: ["string", "null"] }
            timezone: { type: string }
            access:
              type: object
              description: Canonical, read-only access assignment summary. Assignment changes are managed by authorised administrators.
              required: [kind, stateNames, sectors]
              properties:
                kind: { type: string, enum: [organisation_wide, state_scoped, sector_scoped, not_assigned] }
                stateNames: { type: array, items: { type: string } }
                sectors: { type: array, items: { type: string } }
            notificationPreferences: { $ref: "#/components/schemas/NotificationPreferences" }
    UpdateProfileInput:
      description: Only personal contact details and display preferences can be updated here. Identity, access, lifecycle, email, username, and photo paths are server-managed.
      additionalProperties: false
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 150 }
        phone: { type: ["string", "null"], description: "International number, normalised to E.164." }
        jobTitle: { type: ["string", "null"], maxLength: 120 }
        languagePreference: { type: string, enum: [en, ar] }
        timezone: { type: string }
        notificationPreferences:
          oneOf:
            - $ref: "#/components/schemas/NotificationPreferences"
            - type: "null"
    ProfilePhotoUploadRequest:
      type: object
      required: [size, contentType]
      additionalProperties: false
      properties:
        size: { type: integer, minimum: 1, maximum: 5242880 }
        contentType: { type: string, enum: [image/jpeg, image/png, image/webp] }
    ProfilePhotoUploadDescriptor:
      type: object
      required: [uploadURL, uploadToken]
      properties:
        uploadURL: { type: string, format: uri, description: "Short-lived direct upload URL. Never store or display it." }
        uploadToken: { type: string, description: "Short-lived descriptor bound to the signed-in user. Submit only to complete the photo upload." }
    ProfilePhotoCompleteInput:
      type: object
      required: [uploadToken]
      additionalProperties: false
      properties:
        uploadToken: { type: string }
    ProfilePhotoResponse:
      type: object
      required: [avatarUrl]
      properties:
        avatarUrl: { type: ["string", "null"], description: "Authenticated self-profile photo proxy URL; never a storage object path." }
    NotificationPreferences:
      type: object
      description: >
        Validated notification preferences. Nested objects are partial so clients
        may update a subset; unknown keys are rejected. Daily and weekly digest
        modes are intentionally unavailable until a scheduler exists.
      additionalProperties: false
      properties:
        inApp:
          type: object
          additionalProperties: false
          properties:
            approvals: { type: boolean }
            approvalDecisions: { type: boolean }
            comments: { type: boolean }
            assignments: { type: boolean }
            mentions: { type: boolean }
            dueDates: { type: boolean }
            overdueItems: { type: boolean }
            highRisks: { type: boolean }
            criticalRisks: { type: boolean }
            systemNotifications: { type: boolean }
        email:
          type: object
          additionalProperties: false
          properties:
            approvalRequests: { type: boolean }
            approvalDecisions: { type: boolean }
            assignments: { type: boolean }
            mentions: { type: boolean }
            passwordReset: { type: boolean }
            userInvitations: { type: boolean }
            dueDateReminders: { type: boolean }
            highRisks: { type: boolean }
            criticalRisks: { type: boolean }
        deliveryOption:
          type: string
          enum: [inapp_only, email_only, both]
        digest:
          type: string
          enum: [immediate]
        quietHours:
          type: object
          additionalProperties: false
          properties:
            enabled: { type: boolean }
            start: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" }
            end: { type: string, pattern: "^([01]\\d|2[0-3]):[0-5]\\d$" }
            timezone:
              type: string
              enum: [Africa/Khartoum, Africa/Juba, Africa/Cairo, Africa/Nairobi, Africa/Addis_Ababa, Africa/Lagos, Europe/London, Europe/Berlin, Asia/Dubai, America/New_York, America/Los_Angeles, UTC]
...
    ValidationErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        fields:
          type: array
          items:
            type: object
            required: [path, message]
            properties:
              path: { type: string }
              message: { type: string }
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, format: password }
        newPassword: { type: string, format: password }
...